+++ accounts-journey/configuration/Design.swift
@@ -75,14 +75,5 @@ extension AccountsJourney {
         }
         
         public static let darkIconBackgroundColor: UIColor = DesignSystem.shared.colors.surfaceSecondary.default
-        
-        public var errorIconName: ErrorFormatter<String?> = { error in
-            switch error {
-            case .noAccounts:
-                DesignSystem.Assets.icStateViewNoWifi
-            default:
-                DesignSystem.Assets.icStateViewLoadingFailed
-            }
-        }
     }
 }
