+++ accounts-journey/configuration/account-list/AccountsListDesign.swift
@@ -13,5 +13,14 @@ extension AccountsList {
     public struct Design {
         /// Configuration of styles used in AccountList screen
         var styles = Styles()
+        
+        public var errorIconName: ErrorFormatter<String?> = { error in
+            switch error {
+            case .noAccounts:
+                DesignSystem.Assets.icStateViewNoWifi
+            default:
+                DesignSystem.Assets.icStateViewLoadingFailed
+            }
+        }
     }
 }
