--- accounts-journey/presentation/mapper/AccountsJourneyError.swift
@@ -1,13 +1,0 @@
-//
-//  AccountsJourneyError.swift
-//  GoldenSampleApp
-//
-//  Created by Backbase R&D B.V. on 12/10/2023.
-//
-
-import Foundation
-
-public enum AccountsJourneyError: Swift.Error {
-    case invalidResponse
-    case emptyResponse
-}
